HierarchicalLattices.jl Documentation

HierarchicalLattices.DiamondLattice — Type

Main type for a diamond lattice. Takes four fields

generation: The order or generation of the lattice.

initial_state: The initial state of the lattice at time of creation.

final_state: Final state of the lattice after MCMC simulation.

interaction_weight: Strength of the bonds of the lattice.

No type restrictions are imposed on the fields (expect to be changed). There are two external constructors.

DiamondLattice(G::MetaGraph, o::Integer): Construct a diamond lattice with initiial and final states G and o. The strength of the lattice will be G.defaultweight. Note: this will create a deepcopy of G.

DiamondLattice(G::MetaGraph, o::Integer, latticeweight::Number): Construct a diamond lattice with initiial and final states G and o. The strength of the lattice will be latticeweight. Note: this will create a deepcopy of G.

HierarchicalLattices.diamond_ising_lattice — Method

Generates an arbitrary order diamond Ising lattice given an initial state where the initial state can be :zero or :infty. This function does not calculate the location of every spin. Inputs:

order::Int64: Order of the lattice to generate.

b::Int64: Branching number of the lattice.

showprogress (default false): Add a progress bar.

gc_freq (default 0.): Since the function iterates over every edge and adds new nodes with temporary variables at each iteration, for larger lattices this parameter can be helpful to keep memory usage down. Setting this to anything above 0 disables multithreading. At every iteration, run the garbage collector with the probability gc_freq. Any value above 1 will indicate running this at every iteration.

HierarchicalLattices.diamond_ising_lattice — Method

Generates an arbitrary order diamond Ising lattice given an initial state where the initial state can be :zero or :infty. This method will calculate the location of every spin as well for plotting. Inputs:

order::Integer: Order of lattice to generate.

state::Symbol: Can either be :zero for all spins pointing up (+1) or :infty for all spins pointing randomly.

HierarchicalLattices.diamond_lattice — Method

Generates an arbitrary order diamond lattice. This function does not calculate physical locations of every new point. Inputs:

order::Int64: Order of the lattice to create.

b::Int64: Branching number of the lattice.

showprogress (default false): Add a progress bar.

gc_freq (default 0.): Since the function iterates over every edge and adds new nodes with temporary variables at each iteration, for larger lattices this parameter can be helpful to keep memory usage down. Setting this to anything above 0 disables multithreading. At every iteration, run the garbage collector with the probability gc_freq. Any value above 1 will indicate running this at every iteration.

HierarchicalLattices.diamond_lattice — Method

Generates an arbitrary order diamond lattice with b = 2. Thus function will calculate the location of every spin. The location of every new spin will be scaled by 0.5^o where o is the order of the spin. Inputs:

order::Integer: Order of the lattice to generate

HierarchicalLattices.diamond_order_zero_transform! — Method

Takes a lattice graph and an edge and applies b = 2 diamond transform. Thus function will also calculate the physical locations of the new spins with respect to locations of the old one, and thus should only be used for visual demonstration purposes. It is, otherwise, quite memory inefficient. Inputs for the function are:

lattice: The MetaGraph to apply the transformation to.

edge: The edge to apply the transformation to.

scale (default = 1): How much to scale the locations of the new bonds by. A scale of 1 indicates that the distance between the new bond and edge should be equal to the length of edge.

HierarchicalLattices.make_diamond_lattice0 — Method

Creates an order 0, b = 2 diamond lattice with spin values set to false and locations of the spins at 1im and -1im. Returns a MetaGraph.