API Reference

Diamond Lattice Generation

HierarchicalLattices.DiamondLattice — Type

Main type for a diamond lattice. Takes four fields

generation: The order or generation of the lattice.

initial_state: The initial state of the lattice at time of creation.

final_state: Final state of the lattice after MCMC simulation.

interaction_weight: Strength of the bonds of the lattice.

No type restrictions are imposed on the fields (expect to be changed). There are two external constructors.

DiamondLattice(G::MetaGraph, o::Integer): Construct a diamond lattice with initiial and final states G and o. The strength of the lattice will be G.defaultweight. Note: this will create a deepcopy of G.

DiamondLattice(G::MetaGraph, o::Integer, latticeweight::Number): Construct a diamond lattice with initiial and final states G and o. The strength of the lattice will be latticeweight. Note: this will create a deepcopy of G.

HierarchicalLattices.diamond_ising_lattice — Method

Generates an arbitrary order diamond Ising lattice given an initial state where the initial state can be :zero or :infty. This function does not calculate the location of every spin. Inputs:

order::Int64: Order of the lattice to generate.

b::Int64: Branching number of the lattice.

showprogress (default false): Add a progress bar.

gc_freq (default 0.): Since the function iterates over every edge and adds new nodes with temporary variables at each iteration, for larger lattices this parameter can be helpful to keep memory usage down. Setting this to anything above 0 disables multithreading. At every iteration, run the garbage collector with the probability gc_freq. Any value above 1 will indicate running this at every iteration.

HierarchicalLattices.diamond_ising_lattice — Method

Generates an arbitrary order diamond Ising lattice given an initial state where the initial state can be :zero or :infty. This method will calculate the location of every spin as well for plotting. Inputs:

order::Integer: Order of lattice to generate.

state::Symbol: Can either be :zero for all spins pointing up (+1) or :infty for all spins pointing randomly.

HierarchicalLattices.diamond_lattice — Method

Generates an arbitrary order diamond lattice. This function does not calculate physical locations of every new point. Inputs:

order::Int64: Order of the lattice to create.

b::Int64: Branching number of the lattice.

showprogress (default false): Add a progress bar.

gc_freq (default 0.): Since the function iterates over every edge and adds new nodes with temporary variables at each iteration, for larger lattices this parameter can be helpful to keep memory usage down. Setting this to anything above 0 disables multithreading. At every iteration, run the garbage collector with the probability gc_freq. Any value above 1 will indicate running this at every iteration.

HierarchicalLattices.diamond_lattice — Method

Generates an arbitrary order diamond lattice with b = 2. Thus function will calculate the location of every spin. The location of every new spin will be scaled by 0.5^o where o is the order of the spin. Inputs:

order::Integer: Order of the lattice to generate

HierarchicalLattices.diamond_order_zero_transform! — Method

Takes a lattice graph and an edge and applies b = 2 diamond transform. Thus function will also calculate the physical locations of the new spins with respect to locations of the old one, and thus should only be used for visual demonstration purposes. It is, otherwise, quite memory inefficient. Inputs for the function are:

lattice: The MetaGraph to apply the transformation to.

edge: The edge to apply the transformation to.

scale (default = 1): How much to scale the locations of the new bonds by. A scale of 1 indicates that the distance between the new bond and edge should be equal to the length of edge.

HierarchicalLattices.make_diamond_lattice0 — Method

Creates an order 0, b = 2 diamond lattice with spin values set to false and locations of the spins at 1im and -1im. Returns a MetaGraph.

HierarchicalLattices.order_zero_diamond_lattice — Method

Creates an order 0 diamond lattice. Does not calculate the location of the spins.

HierarchicalLattices.raise_order_diamond! — Method

Raise order of diamond lattice. This will apply transform_edge_diamond! to all edges of the lattice, given a b. Inputs:

lattice: MetaGraph to apply the transformation to.

b: Number of new branches created for every edge.

showprogress (default false): Add a progress bar.

gc_freq (default 0.01): Since the function iterates over every edge and adds new nodes with temporary variables at each iteration, for larger lattices this parameter can be helpful to keep memory usage down. Setting this to anything above 0 disables multithreading. At every iteration, run the garbage collector with the probability gc_freq. Any value above 1 will indicate running this at every iteration.

HierarchicalLattices.transform_edge_diamond! — Method

Takes a lattice graph and an edge and applies diamond transform. This will apply the diamond transform with arbitrary b. Inputs:

lattice: MetaGraph lattice to apply the transformation to.

edge: Edge to apply the transformation to.

b: Number of new branches created.

HierarchicalLattices.StackedDiamondLattice — Type

Represents a stacked diamond lattice with anisotropic bond strength in the stacking and "spacial" direction. Fields:

generation: Lattice generation in the spacial direction.

stackingweight: Bond strength in the stacking direction.

initial_state: Initial state at the time of construction.

final_state: Final state after MCMC evolution.

edgeweightmap: Maps vertex index to a dictionary mapping the edge weight to neighbouring spins.

Apart from the inner constructor, this type also has an outer constructor.

StackedDiamondLattice(order, depth, stackingweight, initstate) - Create a stacked diamond lattice with spacial part made of diamond lattice with the given order, depth and stacking weight. Additionally, each spacial lattice in the spacial direction will have an initial state initstate which can be :zero or :infty.

Metropolis MCMC Evolution

HierarchicalLattices.IsingData — Type

Main type used by the metropolis! function. The fields are:

lattice: Lattice to evolve.

magnetization_history::Vector{Float64}: Every index i corresponds to the magnetization at the state i.

internalenergy_history::Vector{Float64}: Every index i corresponds to the internal energy at the state i.

spinflip_history::Vector: Every index i corresponds to the spin flipped in the previous Metropolis step i-1.

HierarchicalLattices.fill_data! — Method

Fill the internal energy or magnetization history of a lattice evolved using the metropolis! function. Inputs:

data::IsingData: Lattice to fill data for.

variable::Symbol: Which variable to fill. Possible values are :U for internal energy and :M for magnetization.

showprogress (default false): Add a progress bar.

progressoutput (default stdout): Write the progress bar output to this IO.

HierarchicalLattices.metropolis! — Method

Evolves the system the given number of steps.

data::IsingData: Lattice to evolve.

steps::Integer: Number of steps to evolve the system for.

T::Float64: Temperature at which to evolve the system.

showprogress (default false): Add a progress bar.

progressoutput (default stdout): Write the progress bar output to this IO.

Cluster (Wolff) MCMC Evolution

HierarchicalLattices.WolffData — Type

Main type used by the wolff! function. Fields:

lattice::T: Lattice to evolve using the Wolff algorithm.

magnetization_history::Vector{Float64}: Every index i corresponds to the magnetization of the system at the wolff step i*saveinterval.

internalenergy_history::Vector{Float64}: Every index i corresponds to the internal energy of the system at the wolff step i*saveinterval.

saveinterval::Int64: Save the state of the system after this many steps.

thermalization_steps::Int64: Steps the system for this many steps before starting data collection.

HierarchicalLattices.wolff! — Method

Run the Wolff algorithm for a given number of steps at a temperature. Inputs:

WD::WolffData{L} WolffData with lattice type L.

nsteps: Number of steps to evolve the system for.

T: Temperature at which to evolve the system.

showprogress (default false): Add a progress bar.

verbose (default false): Look at the argument name.

progressoutput (default stdout): Write the progress output to the given IO.

HierarchicalLattices.wolffstep! — Method

Step the system for one step of the Wolff algorithm. Inputs:

lattice::DiamondLattice: Lattice to evolve. This function is specialized for the diamond lattice.

P_add::Float64: Probability of adding the a spin to the cluster.

adjacentmap::Dictionary{Int64, Vector{Int64}}: Adjacency map of the system. Will be changed to use the adjacency matrix.

HierarchicalLattices.wolffstep! — Method

Step the system for one step of the Wolff algorithm. Inputs:

lattice::StackedDiamondLattice: Lattice to evolve. This function is specialized for the stacked diamond lattice.

probdict::Dictionary{Float64, Float64}: Probability of adding the a spin to the cluster. Maps the bond (or edge) strength to the probability.

adjacentmap::Dictionary{Int64, Vector{Int64}}: Adjacency map of the system. Will be changed to use the adjacency matrix.